backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
sentence-transformers==3.3.1
scikit-learn==1.6.1
PyGithub==2.5.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
pydantic==2.10.4
aiofiles==24.1.0
psycopg2-binary==2.9.10
faiss-cpu==1.9.0.post1
numpy==1.26.4
google-generativeai==0.8.3
openai>=1.0.0
torch
cachetools>=5.3.0
tinydb>=4.8.0
anthropic>=0.34.0
fpdf2